# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Todo API - Built on FL0",
        "description": "A simple todo list API"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "tags": [
        {
            "name": "todo",
            "description": "Todo CRUD operations"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "definitions": {
        "Todo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "completed": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "paths": {
        "/api/todo": {
            "get": {
                "description": "Get all todos",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Todo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new todo",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "body",
                        "description": "Todo object that needs to be added",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Todo"
                        }
                    }
                }
            }
        },
        "/api/todo/{id}": {
            "get": {
                "description": "Get a todo by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Todo id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Todo"
                        }
                    }
                }
            },
            "put": {
                "description": "Toggle a todo completed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Todo id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Todo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a todo by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Todo id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    }
}